seperate files of bar charts and charts become dinamic
