+++ src/components/charts/index.js
@@ -1,0 +1,131 @@
+import React from "react";
+import * as echarts from "echarts/core";
+import {
+    LineChart,
+    BarChart,
+    PieChart,
+    ScatterChart,
+    RadarChart,
+    MapChart,
+    TreeChart,
+    TreemapChart,
+    GraphChart,
+    GaugeChart,
+    FunnelChart,
+    ParallelChart,
+    SankeyChart,
+    BoxplotChart,
+    CandlestickChart,
+    EffectScatterChart,
+    LinesChart,
+    HeatmapChart,
+    PictorialBarChart,
+    ThemeRiverChart,
+    SunburstChart,
+    CustomChart,
+} from "echarts/charts";
+import {
+    GridSimpleComponent,
+    GridComponent,
+    PolarComponent,
+    RadarComponent,
+    GeoComponent,
+    SingleAxisComponent,
+    ParallelComponent,
+    CalendarComponent,
+    GraphicComponent,
+    ToolboxComponent,
+    TooltipComponent,
+    AxisPointerComponent,
+    BrushComponent,
+    TitleComponent,
+    TimelineComponent,
+    MarkPointComponent,
+    MarkLineComponent,
+    MarkAreaComponent,
+    LegendComponent,
+    LegendScrollComponent,
+    LegendPlainComponent,
+    DataZoomComponent,
+    DataZoomInsideComponent,
+    DataZoomSliderComponent,
+    VisualMapComponent,
+    VisualMapContinuousComponent,
+    VisualMapPiecewiseComponent,
+    AriaComponent,
+    TransformComponent,
+} from "echarts/components";
+
+import { CanvasRenderer, SVGRenderer } from "echarts/renderers";
+
+import BarChartComp from "./BarChartComp";
+
+echarts.use([
+    GridSimpleComponent,
+    GridComponent,
+    PolarComponent,
+    RadarComponent,
+    GeoComponent,
+    SingleAxisComponent,
+    ParallelComponent,
+    CalendarComponent,
+    GraphicComponent,
+    ToolboxComponent,
+    TooltipComponent,
+    AxisPointerComponent,
+    BrushComponent,
+    TitleComponent,
+    TimelineComponent,
+    MarkPointComponent,
+    MarkLineComponent,
+    MarkAreaComponent,
+    LegendComponent,
+    LegendScrollComponent,
+    LegendPlainComponent,
+    DataZoomComponent,
+    DataZoomInsideComponent,
+    DataZoomSliderComponent,
+    VisualMapComponent,
+    VisualMapContinuousComponent,
+    VisualMapPiecewiseComponent,
+    AriaComponent,
+    TransformComponent,
+
+    LineChart,
+    BarChart,
+    PieChart,
+    ScatterChart,
+    RadarChart,
+    MapChart,
+    TreeChart,
+    TreemapChart,
+    GraphChart,
+    GaugeChart,
+    FunnelChart,
+    ParallelChart,
+    SankeyChart,
+    BoxplotChart,
+    CandlestickChart,
+    EffectScatterChart,
+    LinesChart,
+    HeatmapChart,
+    PictorialBarChart,
+    ThemeRiverChart,
+    SunburstChart,
+    CustomChart,
+
+    CanvasRenderer,
+    SVGRenderer,
+]);
+
+function NotFound() {
+    return (<h6>Not Found</h6>)
+}
+const ChartTypesComp = {
+    "bar-chart": BarChartComp,
+};
+
+export default function ChartHandler({ type, ...rest }) {
+    const RenderChart = ChartTypesComp[type] !== undefined ? ChartTypesComp[type] : NotFound
+    return <RenderChart {...rest} echarts={echarts} />
+}
